#34 feat: 컴포넌트 재사용 방식으로 잼수정페이지 연결 적용
