+++ client/src/Pages/JamEdit.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import JamMake from './JamMake';
+
+const JamEdit = () => {
+  return <JamMake isEdit />;
+};
+
+export default JamEdit;
